+++ src/routes/index.tsx
@@ -74,6 +74,7 @@ export default function Router() {
         },
         {
           path: "semester",
+          element: <Semester />,
           children: [{ path: "create", element: <CreateSemester /> }],
         },
       ],
@@ -106,6 +107,7 @@ const StudentType = Loadable(
 const CreateStudentType = Loadable(
   lazy(() => import("../pages/Settings/CreateStudentType"))
 );
+const Semester = Loadable(lazy(() => import("../pages/Semester/Semester")));
 const CreateSemester = Loadable(
   lazy(() => import("../pages/Semester/CreateSemester"))
 );